images/child-device/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
paho-mqtt==1.6.1
pipreqs==0.4.11
requests==2.28.1
urllib3==1.26.12
yarg==0.1.9
psutil==5.9.4
